refactor: split text normalization out of util, collect errors, validate the voice index
phoonnx.util had grown into three unrelated things: language-tag identity,
the whole text normalization pipeline, and the per-language lookup tables it
runs on. The normalization pipeline now lives in phoonnx.normalization, one
module per concern — numbers, dates and times, units, and the word-by-word
pass that drives them — with the CONTRACTIONS/TITLES/UNITS tables kept apart
as data, so adding a language is a data change. phoonnx.util re-exports every
name it exported before and keeps the language-tag helpers.

The logger fallback moves to phoonnx.log so the normalization modules can use
it without importing util; phoonnx.util.LOG still resolves for its callers.

phoonnx.errors collects the typed failures and states the taxonomy a caller
needs to act on one: missing language, unsupported language, config defect,
resource refusal. The classes stay defined next to the code that raises them
and are re-exported, so both import paths are the same object.

phoonnx.index_schema reads the voice-index schema off the TTSModelInfo
dataclass rather than restating it, so schema and loader cannot drift. A test
validates every bundled index file: unknown fields, missing required fields,
wrong types, values outside an enum, and unparseable language tags are all
errors. Sequence annotations are described alongside mappings and scalars so
a future List/Tuple/Set field validates rather than crashing, and any
annotation shape the describer does not handle refuses by name instead of
falling through to an isinstance call that cannot work.

Language identity is left in place. scriptconv has no normalize_lang at all,
and its match_lang/_match_lang exist only as BasePhonemizer staticmethods
with a different contract (a str that raises, not a (lang, distance) pair),
so there is no equivalent public surface to re-export and no deprecation to
add.

Fail-before: injecting a "vocoder_ulr" typo into a bundled index file turns
tests/test_index_schema.py from all-green to 2 failed; reverting the sequence
arm turns the new List field tests red with "Subscripted generics cannot be
used with class and instance checks"; dropping the util re-export block turns
tests/test_normalization_package.py from 11 passed to 4 failed with
AttributeError on CONTRACTIONS/TITLES/UNITS. Full non-training suite:
33 failed / 2247 passed against a 33 failed / 2213 passed baseline, identical
failure sets.
