+++ phoonnx/log.py
@@ -1,0 +1,8 @@
+"""Logger used across phoonnx, falling back to the stdlib when ovos-utils is absent."""
+
+try:
+    from ovos_utils.log import LOG
+except ImportError:
+    import logging
+
+    LOG = logging.getLogger("phoonnx")
