+++ phoonnx/normalization/tables.py
@@ -1,0 +1,352 @@
+"""Per-language lookup tables consumed by the normalization pipeline.
+
+Keys are ISO-639 language codes; a locale tag is reduced to its language
+subtag before lookup, so "pt-BR" and "pt-PT" share one table.
+"""
+
+# A dictionary of common contractions and their expanded forms.
+# This list is very comprehensive for English.
+CONTRACTIONS = {
+    "en": {
+        "I'd": "I would",
+        "I'll": "I will",
+        "I'm": "I am",
+        "I've": "I have",
+        "ain't": "is not",
+        "aren't": "are not",
+        "can't": "can not",
+        "could've": "could have",
+        "couldn't": "could not",
+        "didn't": "did not",
+        "doesn't": "does not",
+        "don't": "do not",
+        "gonna": "going to",
+        "gotta": "got to",
+        "hadn't": "had not",
+        "hasn't": "has not",
+        "haven't": "have not",
+        "he'd": "he would",
+        "he'll": "he will",
+        "he's": "he is",
+        "how'd": "how did",
+        "how'll": "how will",
+        "how's": "how is",
+        "isn't": "is not",
+        "it'd": "it would",
+        "it'll": "it will",
+        "it's": "it is",
+        "might've": "might have",
+        "mightn't": "might not",
+        "must've": "must have",
+        "mustn't": "must not",
+        "needn't": "need not",
+        "oughtn't": "ought not",
+        "shan't": "shall not",
+        "she'd": "she would",
+        "she'll": "she will",
+        "she's": "she is",
+        "should've": "should have",
+        "shouldn't": "should not",
+        "somebody's": "somebody is",
+        "someone'd": "someone would",
+        "someone'll": "someone will",
+        "someone's": "someone is",
+        "that'd": "that would",
+        "that'll": "that will",
+        "that's": "that is",
+        "there'd": "there would",
+        "there're": "there are",
+        "there's": "there is",
+        "they'd": "they would",
+        "they'll": "they will",
+        "they're": "they are",
+        "they've": "they have",
+        "wasn't": "was not",
+        "we'd": "we would",
+        "we'll": "we will",
+        "we're": "we are",
+        "we've": "we have",
+        "weren't": "were not",
+        "what'd": "what did",
+        "what'll": "what will",
+        "what're": "what are",
+        "what's": "what is",
+        "what've": "what have",
+        "whats": "what is",
+        "when'd": "when did",
+        "when's": "when is",
+        "where'd": "where did",
+        "where's": "where is",
+        "where've": "where have",
+        "who'd": "who would",
+        "who'd've": "who would have",
+        "who'll": "who will",
+        "who're": "who are",
+        "who's": "who is",
+        "who've": "who have",
+        "why'd": "why did",
+        "why're": "why are",
+        "why's": "why is",
+        "won't": "will not",
+        "won't've": "will not have",
+        "would've": "would have",
+        "wouldn't": "would not",
+        "wouldn't've": "would not have",
+        "y'ain't": "you are not",
+        "y'aint": "you are not",
+        "y'all": "you all",
+        "ya'll": "you all",
+        "you'd": "you would",
+        "you'd've": "you would have",
+        "you'll": "you will",
+        "you're": "you are",
+        "you've": "you have",
+        "I'm'a": "I am going to",
+        "I'm'o": "I am going to",
+        "I'll've": "I will have",
+        "I'd've": "I would have",
+        "Whatcha": "What are you",
+        "amn't": "am not",
+        "'cause": "because",
+        "can't've": "cannot have",
+        "couldn't've": "could not have",
+        "daren't": "dare not",
+        "daresn't": "dare not",
+        "dasn't": "dare not",
+        "everyone's": "everyone is",
+        "gimme": "give me",
+        "gon't": "go not",
+        "hadn't've": "had not have",
+        "he've": "he would have",
+        "he'll've": "he will have",
+        "he'd've": "he would have",
+        "here's": "here is",
+        "how're": "how are",
+        "how'd'y": "how do you do",
+        "howd'y": "how do you do",
+        "howdy": "how do you do",
+        "'tis": "it is",
+        "'twas": "it was",
+        "it'll've": "it will have",
+        "it'd've": "it would have",
+        "kinda": "kind of",
+        "let's": "let us",
+        "ma'am": "madam",
+        "may've": "may have",
+        "mayn't": "may not",
+        "mightn't've": "might not have",
+        "mustn't've": "must not have",
+        "needn't've": "need not have",
+        "ol'": "old",
+        "oughtn't've": "ought not have",
+        "sha'n't": "shall not",
+        "shan't": "shall not",
+        "shalln't": "shall not",
+        "shan't've": "shall not have",
+        "she'd've": "she would have",
+        "shouldn't've": "should not have",
+        "so've": "so have",
+        "so's": "so is",
+        "something's": "something is",
+        "that're": "that are",
+        "that'd've": "that would have",
+        "there'll": "there will",
+        "there'd've": "there would have",
+        "these're": "these are",
+        "they'll've": "they will have",
+        "they'd've": "they would have",
+        "this's": "this is",
+        "this'll": "this will",
+        "this'd": "this would",
+        "those're": "those are",
+        "to've": "to have",
+        "wanna": "want to",
+        "we'll've": "we will have",
+        "we'd've": "we would have",
+        "what'll've": "what will have",
+        "when've": "when have",
+        "where're": "where are",
+        "which's": "which is",
+        "who'll've": "who will have",
+        "why've": "why have",
+        "will've": "will have",
+        "y'all're": "you all are",
+        "y'all've": "you all have",
+        "y'all'd": "you all would",
+        "y'all'd've": "you all would have",
+        "you'll've": "you will have"
+    }
+}
+
+# Dictionaries for titles, units, and their full word equivalents.
+TITLES = {
+    "en": {
+        "Dr.": "Doctor",
+        "Mr.": "Mister",
+        "Prof.": "Professor"
+    },
+    "ca": {
+        "Dr.": "Doctor",
+        "Sr.": "Senyor",
+        "Sra.": "Senyora",
+        "Prof.": "Professor"
+    },
+    "es": {
+        "Dr.": "Doctor",
+        "Sr.": "Señor",
+        "Sra.": "Señora",
+        "Prof.": "Profesor",
+        "D.": "Don",
+        "Dña.": "Doña"
+    },
+    "pt": {
+        "Dr.": "Doutor",
+        "Sr.": "Senhor",
+        "Sra.": "Senhora",
+        "Prof.": "Professor",
+        "Drª.": "Doutora",
+        "Eng.": "Engenheiro",
+        "D.": "Dom",
+        "Dª": "Dona"
+    },
+    "gl": {
+        "Dr.": "Doutor",
+        "Sr.": "Señor",
+        "Sra.": "Señora",
+        "Prof.": "Profesor",
+        "Srta.": "Señorita"
+    },
+    "fr": {
+        "Dr.": "Docteur",
+        "M.": "Monsieur",
+        "Mme": "Madame",
+        "Mlle": "Mademoiselle",
+        "Prof.": "Professeur",
+        "Pr.": "Professeur"
+    },
+    "it": {
+        "Dr.": "Dottore",
+        "Sig.": "Signore",
+        "Sig.ra": "Signora",
+        "Prof.": "Professore",
+        "Dott.ssa": "Dottoressa",
+        "Sig.na": "Signorina"
+    },
+    "nl": {
+        "Dr.": "Dokter",
+        "Dhr.": "De Heer",
+        "Mevr.": "Mevrouw",
+        "Prof.": "Professor",
+        "Drs.": "Dokterandus",
+        "Ing.": "Ingenieur"
+    },
+    "de": {
+        "Dr.": "Doktor",
+        "Prof.": "Professor"
+    }
+}
+
+UNITS = {
+    "en": {
+        "€": "euros",
+        "%": "per cent",
+        "°C": "degrees celsius",
+        "°F": "degrees fahrenheit",
+        "°K": "degrees kelvin",
+        "°": "degrees",
+        "$": "dollars",
+        "£": "pounds",
+        "km": "kilometers",
+        "m": "meters",
+        "cm": "centimeters",
+        "mm": "millimeters",
+        "ft": "feet",
+        "in": "inches",
+        "yd": "yards",
+        "mi": "miles",
+        "kg": "kilograms",
+        "g": "grams",
+        "lb": "pounds",
+        "oz": "ounces",
+        "L": "liters",
+        "mL": "milliliters",
+        "gal": "gallons",
+        "qt": "quarts",
+        "pt": "pints",
+        "hr": "hours",
+        "min": "minutes",
+        "s": "seconds"
+    },
+    "pt": {
+        "€": "euros",
+        "%": "por cento",
+        "°C": "graus celsius",
+        "°F": "graus fahrenheit",
+        "°K": "graus kelvin",
+        "°": "graus",
+        "$": "dólares",
+        "£": "libras",
+        "km": "quilômetros",
+        "m": "metros",
+        "cm": "centímetros",
+        "mm": "milímetros",
+        "kg": "quilogramas",
+        "g": "gramas",
+        "L": "litros",
+        "mL": "mililitros",
+        "h": "horas",
+        "min": "minutos",
+        "s": "segundos"
+    },
+    "es": {
+        "€": "euros",
+        "%": "por ciento",
+        "°C": "grados celsius",
+        "°F": "grados fahrenheit",
+        "°K": "grados kelvin",
+        "°": "grados",
+        "$": "dólares",
+        "£": "libras",
+        "km": "kilómetros",
+        "m": "metros",
+        "cm": "centímetros",
+        "kg": "kilogramos",
+        "g": "gramos",
+        "L": "litros",
+        "mL": "millilitros"
+    },
+    "fr": {
+        "€": "euros",
+        "%": "pour cent",
+        "°C": "degrés celsius",
+        "°F": "degrés fahrenheit",
+        "°K": "degrés kelvin",
+        "°": "degrés",
+        "$": "dollars",
+        "£": "livres",
+        "km": "kilomètres",
+        "m": "mètres",
+        "cm": "centimètres",
+        "kg": "kilogrammes",
+        "g": "grammes",
+        "L": "litres",
+        "mL": "millilitres"
+    },
+    "de": {
+        "€": "Euro",
+        "%": "Prozent",
+        "°C": "Grad Celsius",
+        "°F": "Grad Fahrenheit",
+        "°K": "Grad Kelvin",
+        "°": "Grad",
+        "$": "Dollar",
+        "£": "Pfund",
+        "km": "Kilometer",
+        "m": "Meter",
+        "cm": "Zentimeter",
+        "kg": "Kilogramm",
+        "g": "Gramm",
+        "L": "Liter",
+        "mL": "Milliliter"
+    }
+}
